tests/requirements.txt
# SafeWork Integration Test Requirements

# Core testing framework
pytest==7.4.3
pytest-html==4.1.1
pytest-cov==4.1.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1

# HTTP testing
requests==2.31.0
httpx==0.25.2

# Database testing
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23
redis==5.0.1

# Browser automation (E2E tests)
selenium==4.15.2
webdriver-manager==4.0.1

# Performance monitoring
psutil==5.9.6

# Data processing
pandas==2.0.3
numpy==1.24.3

# Reporting and analysis
Jinja2==3.1.2
matplotlib==3.7.2
seaborn==0.12.2

# Configuration and utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Development tools
black==23.12.1
flake8==7.0.0
mypy==1.7.1

# Security scanning
bandit==1.7.5
safety==2.3.5